requirements.txt
dnspython==2.6.1
et-xmlfile==1.1.0
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
validate_email==1.3
